+++ io/zenoh-links/zenoh-link-unixpipe/src/unix/mod.rs
@@ -24,7 +24,7 @@ pub use unicast::*;
 use zenoh_config::Config;
 use zenoh_core::zconfigurable;
 use zenoh_link_commons::{ConfigurationInspector, LocatorInspector};
-use zenoh_protocol::core::{Locator, ParametersView};
+use zenoh_protocol::core::{Locator, parameters};
 use zenoh_result::ZResult;
 
 pub const UNIXPIPE_LOCATOR_PREFIX: &str = "unixpipe";
@@ -56,7 +56,7 @@ impl ConfigurationInspector<Config> for UnixPipeConfigurator {
             properties.push((config::FILE_ACCESS_MASK, &file_access_mask_));
         }
 
-        let s = ParametersView::from_iter(properties.drain(..));
+        let s = parameters::from_iter(properties.drain(..));
 
         Ok(s)
     }
